Revert "Redesign: CXPO-style overlay panel instead of split tile"
This reverts commit 7019d4ef59356461ea319711a0d56ff3c0a9a998.
