+++ src/features/XIT/LINKEDBUFFERS/LINKEDBUFFERS.vue
@@ -1,7 +1,9 @@
 <script setup lang="ts">
+import TextInput from '@src/components/forms/TextInput.vue';
+import Header from '@src/components/Header.vue';
+import ActionBar from '@src/components/ActionBar.vue';
 import PrunButton from '@src/components/PrunButton.vue';
 import PresetList from './PresetList.vue';
-import ControlPanel from './ControlPanel.vue';
 import { useXitParameters } from '@src/hooks/use-xit-parameters';
 import { useTile } from '@src/hooks/use-tile';
 import { userData } from '@src/store/user-data';
@@ -13,6 +15,10 @@ import { dispatchClientPrunMessage } from '@src/infrastructure/prun-api/prun-api
 import { clickElement, changeInputValue } from '@src/util';
 import { sleep } from '@src/utils/sleep';
 import { isEmpty } from 'ts-extras';
+import { vDraggable } from 'vue-draggable-plus';
+import { grip } from '@src/components/grip';
+import GripCell from '@src/components/grip/GripCell.vue';
+import GripHeaderCell from '@src/components/grip/GripHeaderCell.vue';
 
 const tile = useTile();
 const parameters = useXitParameters();
@@ -31,130 +37,56 @@ const preset = computed(() => {
 
 const inputText = ref('');
 const isProcessing = ref(false);
+const edit = ref(false);
 const childTiles = ref<HTMLElement[]>([]);
 const gridReady = ref(false);
-const panelMounted = ref(false);
 
-// On first open (solo buffer), resize the window and convert the XIT tile
-// into the first child tile by changing its command.
+// Split immediately if this is a solo buffer.
 const isSoloBuffer = tile.container.classList.contains(C.Window.body);
-const goingToSetup = ref(false);
+const goingToSplit = ref(false);
 
 if (isSoloBuffer && preset.value && preset.value.commands.length > 0) {
-  goingToSetup.value = true;
+  goingToSplit.value = true;
   const saved = preset.value.lastBufferSize;
   if (saved) {
     setBufferSize(tile.id, saved[0], saved[1]);
   } else {
     const width = parseInt(tile.container.style.width.replace('px', ''), 10);
     const height = parseInt(tile.container.style.height.replace('px', ''), 10);
-    setBufferSize(tile.id, Math.max(width, 900), Math.max(height, 500));
+    setBufferSize(tile.id, width + 500, height);
   }
+  const splitButton = _$$(tile.frame, C.TileControls.control).find(x => x.textContent === '|');
+  void clickElement(splitButton);
 }
 
+// After re-mount (post-split), build the grid.
 onMounted(async () => {
-  if (!preset.value || preset.value.commands.length === 0) {
+  if (goingToSplit.value || !preset.value) {
     return;
   }
 
-  if (goingToSetup.value) {
-    // First mount: split this tile to create the child grid, then mount
-    // the control panel as an overlay on the window.
-    await setupChildGrid();
+  const rightSibling = getRightSibling();
+  if (!rightSibling) {
+    gridReady.value = true;
     return;
   }
 
-  // If we're re-mounted after a split, just find child tiles.
-  // This handles the case where the XIT tile persists after splits.
-  await findAndTrackChildTiles();
+  const commandCount = preset.value.commands.length;
+  if (commandCount > 1) {
+    await buildGrid(rightSibling, commandCount);
+  }
+
+  childTiles.value = _$$(rightSibling, C.Tile.tile) as HTMLElement[];
+  gridReady.value = true;
 });
 
+// Save window size when the component unmounts (captures user's resize).
 onBeforeUnmount(() => {
   saveBufferSize();
 });
 
-async function setupChildGrid() {
-  const commandCount = preset.value!.commands.length;
-
-  if (commandCount === 1) {
-    // Single command: just change this tile's command after mounting the panel.
-    await mountOverlayPanel();
-    gridReady.value = true;
-    return;
-  }
-
-  // For multiple commands, split this tile to create the grid.
-  // First split creates two tiles from the current one.
-  await splitTileElement(tile.frame.parentElement as HTMLElement, '\u2013');
-
-  // The XIT tile may have been destroyed by the split. Find all tiles in the window.
-  const windowBody = getWindowBody();
-  if (!windowBody) {
-    return;
-  }
-
-  // Continue splitting to create the full grid.
-  await buildGridInWindow(windowBody, commandCount);
-
-  // Mount the overlay panel on the window.
-  await mountOverlayPanel();
-
-  // Find all child tiles.
-  childTiles.value = _$$(windowBody, C.Tile.tile) as HTMLElement[];
-  gridReady.value = true;
-}
-
-async function findAndTrackChildTiles() {
-  const windowBody = getWindowBody();
-  if (!windowBody) {
-    gridReady.value = true;
-    return;
-  }
-
-  childTiles.value = _$$(windowBody, C.Tile.tile) as HTMLElement[];
-  gridReady.value = true;
-
-  // Mount overlay if not already.
-  if (!panelMounted.value) {
-    await mountOverlayPanel();
-  }
-}
-
-function getWindowBody(): HTMLElement | null {
-  const win = tile.frame.closest(`.${C.Window.window}`);
-  if (!win) {
-    return null;
-  }
-  return _$(win as HTMLElement, C.Window.body) as HTMLElement | null;
-}
-
 function getWindowElement(): HTMLElement | null {
   return tile.frame.closest(`.${C.Window.window}`) as HTMLElement | null;
-}
-
-async function mountOverlayPanel() {
-  const win = getWindowElement();
-  if (!win || panelMounted.value) {
-    return;
-  }
-
-  // Make the window a positioning context for the overlay.
-  const body = _$(win, C.Window.body) as HTMLElement;
-  if (!body) {
-    return;
-  }
-  body.style.position = 'relative';
-
-  createFragmentApp(
-    ControlPanel,
-    reactive({
-      preset: preset.value!,
-      inputText,
-      onCommandClick: handleCommandClick,
-    }),
-  ).appendTo(body);
-
-  panelMounted.value = true;
 }
 
 function saveBufferSize() {
@@ -172,6 +104,18 @@ function saveBufferSize() {
   }
 }
 
+function getRightSibling(): HTMLElement | undefined {
+  const isInNodeChild = tile.container.classList.contains(C.Node.child);
+  const isInNode = tile.container.parentElement?.classList.contains(C.Node.node);
+  const isInWindow = tile.container.parentElement?.parentElement?.classList.contains(C.Window.body);
+  if (!isInNodeChild || !isInNode || !isInWindow) {
+    return undefined;
+  }
+
+  const node = tile.container.parentElement!;
+  return _$$(node, C.Node.child).find(x => x !== tile.container) as HTMLElement | undefined;
+}
+
 async function splitTileElement(tileEl: HTMLElement, direction: '|' | '\u2013') {
   const button = _$$(tileEl, C.TileControls.control).find(x => x.textContent === direction);
   if (!button) {
@@ -181,7 +125,7 @@ async function splitTileElement(tileEl: HTMLElement, direction: '|' | '\u2013') 
   await sleep(200);
 }
 
-async function buildGridInWindow(windowBody: HTMLElement, count: number) {
+async function buildGrid(container: HTMLElement, count: number) {
   if (count <= 1) {
     return;
   }
@@ -189,13 +133,11 @@ async function buildGridInWindow(windowBody: HTMLElement, count: number) {
   const rows = Math.ceil(count / 2);
 
   // Create rows by splitting the last tile vertically.
-  // Start from the current number of tiles (may already be 2 from the initial split).
-  let currentTileCount = (_$$(windowBody, C.Tile.tile) as HTMLElement[]).length;
-  while (currentTileCount < rows) {
-    const tiles = _$$(windowBody, C.Tile.tile) as HTMLElement[];
+  for (let i = 1; i < rows; i++) {
+    const tiles = _$$(container, C.Tile.tile) as HTMLElement[];
     const lastTile = tiles[tiles.length - 1];
+    // En-dash for vertical split.
     await splitTileElement(lastTile, '\u2013');
-    currentTileCount = (_$$(windowBody, C.Tile.tile) as HTMLElement[]).length;
   }
 
   if (count <= rows) {
@@ -203,16 +145,21 @@ async function buildGridInWindow(windowBody: HTMLElement, count: number) {
   }
 
   // Split rows that need 2 columns, from the last row upward.
-  const currentTiles = _$$(windowBody, C.Tile.tile) as HTMLElement[];
-  let splitsNeeded = count - currentTiles.length;
+  const currentTiles = _$$(container, C.Tile.tile) as HTMLElement[];
+  let splitsNeeded = count - rows;
   for (let i = currentTiles.length - 1; i >= 0 && splitsNeeded > 0; i--) {
     await splitTileElement(currentTiles[i], '|');
     splitsNeeded--;
   }
 }
 
+function resolveCommand(template: string) {
+  return template.replace(/\{input\}/g, inputText.value.trim()).toUpperCase();
+}
+
 async function changeTileCommand(tileEl: HTMLElement, command: string) {
   const id = getPrunId(tileEl)!;
+  // Clear current command.
   let message = UI_TILES_CHANGE_COMMAND(id, null);
   if (!dispatchClientPrunMessage(message)) {
     const changeButton = _$$(tileEl, C.TileControls.control).find(x => x.textContent === ':');
@@ -220,6 +167,7 @@ async function changeTileCommand(tileEl: HTMLElement, command: string) {
   } else {
     await sleep(0);
   }
+  // Set new command.
   message = UI_TILES_CHANGE_COMMAND(id, command);
   if (!dispatchClientPrunMessage(message)) {
     const input = (await $(tileEl, C.PanelSelector.input)) as HTMLInputElement;
@@ -229,21 +177,15 @@ async function changeTileCommand(tileEl: HTMLElement, command: string) {
   await $(tileEl, C.TileFrame.frame);
 }
 
-async function handleCommandClick(cmd: UserData.LinkedBuffersCommand, index: number) {
-  if (!inputText.value.trim() || isProcessing.value) {
+async function onCommandClick(cmd: UserData.LinkedBuffersCommand, index: number) {
+  if (!inputText.value.trim() || isProcessing.value || edit.value) {
     return;
   }
 
-  const resolved = cmd.template.replace(/\{input\}/g, inputText.value.trim()).toUpperCase();
+  const resolved = resolveCommand(cmd.template);
   isProcessing.value = true;
 
   try {
-    // Re-find child tiles in case DOM changed.
-    const windowBody = getWindowBody();
-    if (windowBody) {
-      childTiles.value = _$$(windowBody, C.Tile.tile) as HTMLElement[];
-    }
-
     const child = childTiles.value[index];
     if (child?.isConnected) {
       await changeTileCommand(child, resolved);
@@ -251,6 +193,24 @@ async function handleCommandClick(cmd: UserData.LinkedBuffersCommand, index: num
   } finally {
     isProcessing.value = false;
   }
+}
+
+function addCommand() {
+  if (!preset.value) {
+    return;
+  }
+  preset.value.commands.push({
+    id: createId(),
+    label: 'CMD',
+    template: 'CMD {input}',
+  });
+}
+
+function deleteCommand(cmd: UserData.LinkedBuffersCommand) {
+  if (!preset.value) {
+    return;
+  }
+  preset.value.commands = preset.value.commands.filter(x => x !== cmd);
 }
 
 function onCreateClick() {
@@ -267,14 +227,98 @@ function onCreateClick() {
 </script>
 
 <template>
-  <div v-if="goingToSetup" />
+  <div v-if="goingToSplit" />
   <PresetList v-else-if="isEmpty(parameters)" />
   <div v-else-if="!preset" :class="$style.create">
     <span>Preset "{{ parameters.join(' ') }}" not found.</span>
     <PrunButton primary :class="$style.createButton" @click="onCreateClick">CREATE</PrunButton>
   </div>
   <div v-else :class="$style.root">
-    <div v-if="!gridReady" :class="$style.status">Setting up...</div>
+    <Header>{{ preset.name }}</Header>
+    <div :class="$style.inputSection">
+      <label :class="$style.label">Input</label>
+      <input
+        v-model="inputText"
+        type="text"
+        :class="$style.input"
+        autocomplete="off"
+        data-1p-ignore="true"
+        data-lpignore="true" />
+    </div>
+    <div v-if="!gridReady" :class="$style.status">Setting up grid...</div>
+    <template v-if="!edit">
+      <table>
+        <thead>
+          <tr>
+            <th>Commands</th>
+          </tr>
+        </thead>
+        <tbody>
+          <tr v-if="preset.commands.length === 0">
+            <td>No commands. Click EDIT to add some.</td>
+          </tr>
+          <template v-else>
+            <tr
+              v-for="(cmd, index) in preset.commands"
+              :key="cmd.id"
+              @click="onCommandClick(cmd, index)">
+              <td :class="$style.commandCell">
+                <span :class="[C.Link.link, $style.commandLink]">
+                  {{ resolveCommand(cmd.template) || cmd.template }}
+                </span>
+              </td>
+            </tr>
+          </template>
+        </tbody>
+      </table>
+      <ActionBar>
+        <PrunButton primary @click="edit = true">EDIT</PrunButton>
+      </ActionBar>
+    </template>
+    <template v-else>
+      <table>
+        <thead>
+          <tr>
+            <GripHeaderCell />
+            <th>Label</th>
+            <th>Template</th>
+            <th />
+          </tr>
+        </thead>
+        <template v-if="preset.commands.length === 0">
+          <tbody>
+            <tr>
+              <td>No commands.</td>
+            </tr>
+          </tbody>
+        </template>
+        <template v-else>
+          <tbody v-draggable="[preset.commands, grip.draggable]">
+            <tr v-for="cmd in preset.commands" :key="cmd.id">
+              <GripCell />
+              <td>
+                <div :class="[C.forms.input, $style.inline]">
+                  <TextInput v-model="cmd.label" />
+                </div>
+              </td>
+              <td>
+                <div :class="C.forms.input">
+                  <TextInput v-model="cmd.template" />
+                </div>
+              </td>
+              <td>
+                <PrunButton danger @click="deleteCommand(cmd)">DELETE</PrunButton>
+              </td>
+            </tr>
+          </tbody>
+        </template>
+      </table>
+      <ActionBar>
+        <PrunButton primary @click="addCommand">ADD COMMAND</PrunButton>
+        <PrunButton primary @click="edit = false">DONE</PrunButton>
+      </ActionBar>
+      <div :class="$style.hint">Close and reopen buffer to apply layout changes.</div>
+    </template>
   </div>
 </template>
 
@@ -282,11 +326,49 @@ function onCreateClick() {
 .root {
   height: 100%;
   display: flex;
+  flex-direction: column;
+  padding: 4px;
+  min-width: 0;
+  overflow: hidden;
+}
+
+.inputSection {
+  display: flex;
   align-items: center;
-  justify-content: center;
+  margin-bottom: 4px;
+  gap: 4px;
+}
+
+.label {
+  white-space: nowrap;
+  font-size: 0.9em;
+}
+
+.input {
+  width: 80px;
+  max-width: 80px;
+  text-transform: uppercase;
+  text-align: left;
+  background: #1a1f22;
+  border: 1px solid #2b485a;
+  color: #bfbfbf;
+  padding: 2px 4px;
+  font-family: inherit;
+  font-size: inherit;
+}
+
+.commandCell {
+  cursor: pointer;
+  padding: 1px 4px;
+  white-space: nowrap;
+}
+
+.commandLink {
+  cursor: pointer;
 }
 
 .status {
+  padding: 4px;
   opacity: 0.7;
 }
 
@@ -301,5 +383,15 @@ function onCreateClick() {
 
 .createButton {
   margin-top: 5px;
+}
+
+.inline {
+  display: inline-block;
+}
+
+.hint {
+  padding: 4px;
+  opacity: 0.5;
+  font-size: 0.9em;
 }
 </style>
